+++ src/libraries/System.Threading.RateLimiting/tests/System.Threading.RateLimiting.Tests.csproj
@@ -4,12 +4,14 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="BaseRateLimiterTests.cs" />
+    <Compile Include="ChainedLimiterTests.cs" />
     <Compile Include="ConcurrencyLimiterTests.cs" />
     <Compile Include="PartitionedRateLimiterTests.cs" />
     <Compile Include="FixedWindowRateLimiterTests.cs" />
     <Compile Include="RateLimiterPartitionTests.cs" />
     <Compile Include="SlidingWindowRateLimiterTests.cs" />
     <Compile Include="TokenBucketRateLimiterTests.cs" />
+    <Compile Include="Infrastructure\Utils.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\src\System.Threading.RateLimiting.csproj" />
